requirements.txt
faiss-cpu
sentence-transformers
langchain-community
langchain
pypdf
streamlit
google-generativeai
from pypdf import PdfReader
from pdfextractor import text_extractor_pdf